+++ lib/MIP/Recipes/Install/Singularity.pm
@@ -20,7 +20,7 @@ use Readonly;
 ## MIPs lib/
 use MIP::Check::Path qw{ check_future_filesystem_for_directory };
 use MIP::Constants
-  qw{ $AT $DOLLAR_SIGN $DOUBLE_QUOTE $COLON $LOG_NAME $NEWLINE $SINGLE_QUOTE $SPACE };
+  qw{ $AT $DOLLAR_SIGN $DOUBLE_QUOTE $COLON $EMPTY_STR $LOG_NAME $NEWLINE $SINGLE_QUOTE $SPACE };
 use MIP::Gnu::Coreutils qw{ gnu_chmod gnu_mkdir gnu_echo };
 use MIP::Language::Shell qw{ build_shebang };
 use MIP::Log::MIP_log4perl qw{ retrieve_log };
@@ -247,6 +247,11 @@ sub setup_singularity_executable {
             $container_executable = $executable_href->{$executable};
         }
 
+        if (    $executable_href->{$executable}
+            and $executable_href->{$executable} eq q{no_executable_in_image} )
+        {
+            $container_executable = $EMPTY_STR;
+        }
         my @singularity_cmds = singularity_exec(
             {
                 singularity_container          => $container_path,