+++ t/manta_workflow.t
@@ -25,7 +25,7 @@ use MIP::Test::Commands qw{ test_function };
 use MIP::Test::Fixtures qw{ test_standard_cli };
 
 my $VERBOSE = 1;
-our $VERSION = 1.00;
+our $VERSION = 1.01;
 
 $VERBOSE = test_standard_cli(
     {
@@ -61,7 +61,7 @@ diag(   q{Test manta_workflow from Manta v}
       . $EXECUTABLE_NAME );
 
 ## Base arguments
-my $function_base_command = q{runWorkflow.py};
+my @function_base_commands = qw{ runWorkflow.py };
 
 my %base_argument = (
     stderrfile_path => {
@@ -74,7 +74,7 @@ my %base_argument = (
     },
     FILEHANDLE => {
         input           => undef,
-        expected_output => $function_base_command,
+        expected_output => \@function_base_commands,
     },
 );
 
@@ -83,7 +83,7 @@ my %base_argument = (
 my %required_argument = (
     outdirectory_path => {
         input           => q{outdirectory_path},
-        expected_output => q{outdirectory_path},
+        expected_output => catfile( q{outdirectory_path}, q{runWorkflow.py} ),
     },
 );
 
@@ -110,14 +110,9 @@ foreach my $argument_href (@arguments) {
         {
             argument_href              => $argument_href,
             do_test_base_command       => 1,
-            function_base_commands_ref => [
-                catfile(
-                    $required_argument{outdirectory_path}{expected_output},
-                    $function_base_command
-                )
-            ],
-            module_function_cref   => $module_function_cref,
-            required_argument_href => \%required_argument,
+            function_base_commands_ref => \@function_base_commands,
+            module_function_cref       => $module_function_cref,
+            required_argument_href     => \%required_argument,
         }
     );
 }